pyproject.toml
[project]
name = "openai-sdk-guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "certifi>=2025.4.26",
    "dnspython>=2.7.0",
    "fastapi>=0.115.12",
    "openai>=1.79.0",
    "openai-agents>=0.0.15",
    "prompt-toolkit>=3.0.51",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.4",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "pyttsx3>=2.98",
    "speechrecognition>=3.14.3",
    "tavily-python>=0.7.3",
    "uvicorn>=0.34.2",
]
